post save and pick

+++ src/SocialAppGeneral/Like.java
@@ -4,13 +4,13 @@ package SocialAppGeneral;
  * Created by khaled hesham on 12/2/2016.
  */
 public class Like extends PostAtachmment {
-    private short like;
+    private int like;
 
-    public short getLike() {
+    public int getLike() {
         return like;
     }
 
-    public void setLike(short like) {//put 1 if put like put -1 if put dislike
+    public void setLike(int like) {//put 1 if put like put -1 if put dislike
         this.like = like;
     }
 }
